backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.46
asyncpg==0.30.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.7.1
